Organize chat component
