+++ app/components/header/Feedback.tsx
@@ -2,7 +2,7 @@ import { toast } from 'react-toastify';
 import ReactModal from 'react-modal';
 import { useState } from 'react';
 import { supabaseSubmitFeedback } from '~/lib/supabase/feedback';
-import { getLastChatMessages } from '~/components/chat/Chat.client';
+import { getLastChatMessages } from '~/utils/chat/messageUtils';
 
 ReactModal.setAppElement('#root');
