+++ app/hooks/useSpeechRecognition.ts
@@ -1,0 +1,67 @@
+import { useState, useEffect } from 'react';
+
+interface UseSpeechRecognitionProps {
+  onTranscriptChange: (transcript: string) => void;
+}
+
+export const useSpeechRecognition = ({ onTranscriptChange }: UseSpeechRecognitionProps) => {
+  const [isListening, setIsListening] = useState(false);
+  const [recognition, setRecognition] = useState<SpeechRecognition | null>(null);
+  const [transcript, setTranscript] = useState('');
+
+  useEffect(() => {
+    if (typeof window !== 'undefined' && ('SpeechRecognition' in window || 'webkitSpeechRecognition' in window)) {
+      const SpeechRecognition = window.SpeechRecognition || window.webkitSpeechRecognition;
+      const recognition = new SpeechRecognition();
+      recognition.continuous = true;
+      recognition.interimResults = true;
+
+      recognition.onresult = (event) => {
+        const transcript = Array.from(event.results)
+          .map((result) => result[0])
+          .map((result) => result.transcript)
+          .join('');
+
+        setTranscript(transcript);
+        onTranscriptChange(transcript);
+      };
+
+      recognition.onerror = (event) => {
+        console.error('Speech recognition error:', event.error);
+        setIsListening(false);
+      };
+
+      setRecognition(recognition);
+    }
+  }, [onTranscriptChange]);
+
+  const startListening = () => {
+    if (recognition) {
+      recognition.start();
+      setIsListening(true);
+    }
+  };
+
+  const stopListening = () => {
+    if (recognition) {
+      recognition.stop();
+      setIsListening(false);
+    }
+  };
+
+  const abortListening = () => {
+    if (recognition) {
+      recognition.abort();
+      setTranscript('');
+      setIsListening(false);
+    }
+  };
+
+  return {
+    isListening,
+    transcript,
+    startListening,
+    stopListening,
+    abortListening,
+  };
+};
