+++ app/routes/_index.tsx
@@ -2,7 +2,7 @@ import { json, type MetaFunction } from '~/lib/remix-types';
 import { Suspense } from 'react';
 import { ClientOnly } from 'remix-utils/client-only';
 import { BaseChat } from '~/components/chat/BaseChat/BaseChat';
-import { Chat } from '~/components/chat/Chat/Chat.client';
+import { Chat } from '~/components/chat/ChatComponent/Chat.client';
 import { PageContainer } from '~/layout/PageContainer';
 export const meta: MetaFunction = () => {
   return [{ title: 'Nut' }];
